app/requirements.txt
streamlit==1.32.2
boto3==1.34.59
psycopg2-binary==2.9.9
pandas==2.2.1
